requirements.txt
Pyrogram==2.0.106
TgCrypto==1.2.3
python-dotenv
aiohttp
pymongo[srv]==3.12.3
motor==2.5.1
